+++ src/translator/table/primary.rs
@@ -1,14 +1,10 @@
-use std::{
-    collections::{HashMap, HashSet},
-    fs,
-};
+use std::collections::{HashMap, HashSet};
 
 use log::warn;
 
 use crate::{
     Direction,
     emphasis::EmphasisSpan,
-    hyphenation::HyphenationTable,
     parser::{
         AnchoredRule, Braille, BrailleChars, CharacterClass, CharacterClasses, HasNocross,
         HasPrecedence, WithClass, fallback,
@@ -1343,11 +1339,8 @@ impl PrimaryTable {
                         ctx,
                     )?;
                 }
-                Rule::IncludeHyphenation { path } => {
-                    let source = fs::read_to_string(path)?;
-                    builder
-                        .hyphenation_constrainer
-                        .hyphenator(HyphenationTable::parse(&source)?);
+                Rule::IncludeHyphenation { table, .. } => {
+                    builder.hyphenation_constrainer.hyphenator(table.clone());
                 }
                 _ => (),
             }